+++ lib/elastic_apm/agent.rb
@@ -49,12 +49,14 @@ module ElasticAPM
     def initialize(config)
       @config = config
 
-      @transport = Transport::Base.new(config)
-      @instrumenter = Instrumenter.new(config) { |event| enqueue event }
-
       @stacktrace_builder = StacktraceBuilder.new(config)
       @context_builder = ContextBuilder.new(config)
       @error_builder = ErrorBuilder.new(self)
+
+      @transport = Transport::Base.new(config)
+      @instrumenter = Instrumenter.new(
+        config, stacktrace_builder: stacktrace_builder
+      ) { |event| enqueue event }
     end
 
     attr_reader :config, :transport, :instrumenter,