+++ lib/elastic_apm/span.rb
@@ -22,20 +22,23 @@ module ElasticAPM
       @name = name
       @type = type || DEFAULT_TYPE
 
-      @id = SecureRandom.hex(8)
-      @parent_id = parent_id
       @transaction_id = transaction_id
 
-      @trace_context = trace_context&.child(self)
+      @parent_id = parent_id
+      @trace_context = trace_context || TraceContext.for_span
 
-      @context = context
+      @context = context || Span::Context.new
       @stacktrace_builder = stacktrace_builder
     end
     # rubocop:enable Metrics/ParameterLists
 
     attr_accessor :name, :type, :original_backtrace, :parent_id, :trace_context
-    attr_reader :id, :context, :stacktrace, :duration,
+    attr_reader :context, :stacktrace, :duration,
       :timestamp, :transaction_id
+
+    def id
+      trace_context&.span_id
+    end
 
     def trace_id
       trace_context&.trace_id
@@ -43,18 +46,18 @@ module ElasticAPM
 
     # life cycle
 
-    def start
-      @timestamp = Util.micros
+    def start(timestamp = Util.micros)
+      @timestamp = timestamp
 
       self
     end
 
-    def stop
-      @duration = Util.micros - timestamp
+    def stop(end_timestamp = Util.micros)
+      @duration ||= (end_timestamp - timestamp)
     end
 
-    def done
-      stop
+    def done(end_time: Util.micros)
+      stop end_time
 
       if should_build_stacktrace?
         build_stacktrace