+++ CHANGELOG.md
@@ -4,18 +4,17 @@ All notable changes to this project will be documented in this file.
 
 This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## 2.2.0 (2019-01-22)
 
 ### Added
 
 - Support for [OpenTracing](https://opentracing.io) ([#273](https://github.com/elastic/apm-agent-ruby/pull/273))
 - Add capture_* options ([#279](https://github.com/elastic/apm-agent-ruby/pull/279))
-- Support ERB tags in config file ([#286](https://github.com/elastic/apm-agent-ruby/pull/286))
+- Evaluate the config file as ERB ([#288](https://github.com/elastic/apm-agent-ruby/pull/288))
 
 ### Changed
 
 - Rename `Traceparent` object to `TraceContext` ([#271](https://github.com/elastic/apm-agent-ruby/pull/271))
-- Evaluate the config file as ERB ([#288](https://github.com/elastic/apm-agent-ruby/pull/288))
 
 ### Fixed